Updated add accommodation to contain logic
