+++ client/src/components/Dashboard/accommodation/accommodation.js
@@ -8,24 +8,6 @@ export const ViewAccommodation = ({ session }) => {
   const userId = session;
   const [accommodation, setAccommodation] = useState([]);
   const [open, setOpen] = useState(false);
-  const [accommodationArray, setAccommodationArray] = useState({
-    name: "",
-    contactNumber: "",
-    checkInDate: "",
-    checkOutDate: "",
-    checkInTime: "",
-    checkOutTime: "",
-    bookingReference: "",
-    buildingNumber: "",
-    buildingName: "",
-    addressLine1: "",
-    addressLine2: "",
-    city: "",
-    postalCode: "",
-    stateCounty: "",
-    countryCode: "",
-    user: userId,
-  });
 
   useEffect(() => {
     if (userId !== "null") {
@@ -35,7 +17,7 @@ export const ViewAccommodation = ({ session }) => {
           setAccommodation(res.data.accommodation);
         });
     }
-  }, [accommodationArray]);
+  }, []);
 
   const handleOpen = () => {
     setOpen(true);
@@ -43,81 +25,6 @@ export const ViewAccommodation = ({ session }) => {
 
   const handleClose = () => {
     setOpen(false);
-  };
-
-  const handleChange = (e) => {
-    const value = e.target.value;
-    setAccommodationArray({
-      ...accommodationArray,
-      [e.target.name]: value,
-    });
-  };
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-
-    const {
-      name,
-      contactNumber,
-      checkInDate,
-      checkOutDate,
-      checkInTime,
-      checkOutTime,
-      bookingReference,
-      buildingNumber,
-      buildingName,
-      addressLine1,
-      addressLine2,
-      city,
-      postalCode,
-      stateCounty,
-      countryCode,
-      user,
-    } = accommodationArray;
-
-    const newAccommodation = {
-      name,
-      contactNumber,
-      checkInDate,
-      checkOutDate,
-      checkInTime,
-      checkOutTime,
-      bookingReference,
-      buildingNumber,
-      buildingName,
-      addressLine1,
-      addressLine2,
-      city,
-      postalCode,
-      stateCounty,
-      countryCode,
-      user,
-    };
-
-    await axios
-      .post("http://localhost:8000/dashboard/accommodation", newAccommodation)
-      .catch((err) => console.log(err.message))
-      .then(() => {
-        setAccommodationArray({
-          name: "",
-          contactNumber: "",
-          checkInDate: "",
-          checkOutDate: "",
-          checkInTime: "",
-          checkOutTime: "",
-          bookingReference: "",
-          buildingNumber: "",
-          buildingName: "",
-          addressLine1: "",
-          addressLine2: "",
-          city: "",
-          postalCode: "",
-          stateCounty: "",
-          countryCode: "",
-          user: userId,
-        });
-        handleClose();
-      });
   };
 
   if (accommodation.length) {
@@ -130,13 +37,13 @@ export const ViewAccommodation = ({ session }) => {
             handleOpen={handleOpen}
             open={open}
             handleClose={handleClose}
-            handleChange={handleChange}
-            accommodation={accommodationArray}
-            handleSubmit={handleSubmit}
+            accommodationData={accommodation}
+            accommodationId={null}
+            userId={userId}
           />
         </div>
         <div className="body">
-          <AccommodationCard accommodation={accommodation} />
+          <AccommodationCard accommodation={accommodation} userId={userId} />
         </div>
       </div>
     );
@@ -150,9 +57,9 @@ export const ViewAccommodation = ({ session }) => {
             handleOpen={handleOpen}
             open={open}
             handleClose={handleClose}
-            handleChange={handleChange}
-            accommodation={accommodationArray}
-            handleSubmit={handleSubmit}
+            accommodationData={accommodation}
+            accommodationId={null}
+            userId={userId}
           />
         </div>
       </div>
